Notify on hosted zone update
This allows configuring an EventBridge Event Rule to notify any time the
created user modifies the specified hosted zone. This includes a link to
the CloudTrail event in the email as well as the actual list of changes
that were made. The old values are not provided for simplicity. The
CloudTrail link should work in any partition.

The Lambda function, for fun, runs specifically on ARM64 but it would
totally work on x86_64 as well.